requirements.txt
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
pyyaml>=5.4.0
gym>=0.21.0
tqdm>=4.62.0
simpy>=4.0.0
scipy>=1.7.0
tensorboard>=2.8.0
pytest>=6.2.0 